# marketing_service_backend/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Marketing Management Service
  description: "A service for adding carmodels, offers and prices for cars"
  contact:
    name: Hari prasath
    email: hariprasath.narayanasamy@gmail.com
  version: 1.0.0
servers:
- url: http://127.0.0.1:7000/marketingservice/v1
tags:
- name: Car Models
- name: Offers
- name: Prices
paths:
  /cars/models:
    get:
      tags:
      - Car Models
      description: Get details of all car models
      operationId: get_car_models
      responses:
        "200":
          description: Fetches all car models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/carmodeldetails'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404NotFoundResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
        default:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/400BadRequestResponse'
      x-openapi-router-controller: swagger_server.controllers.car_models_controller
    post:
      tags:
      - Car Models
      description: Add one car model
      operationId: add_car_model
      requestBody:
        description: Adds a car model
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/carmodelinfo'
      responses:
        "201":
          description: car model added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/201CarModelCreatedResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404NotFoundResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
        default:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/400BadRequestResponse'
      x-openapi-router-controller: swagger_server.controllers.car_models_controller
  /cars/models/{model_type}:
    get:
      tags:
      - Car Models
      description: Get details of all car models
      operationId: get_car_modeltype
      parameters:
      - name: model_type
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Fetches all car models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/carmodeldetails'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/409ConflictResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
      x-openapi-router-controller: swagger_server.controllers.car_models_controller
  /offers:
    get:
      tags:
      - Offers
      description: Get offers
      operationId: get_offer_details
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/offer_details'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/409ConflictResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
      x-openapi-router-controller: swagger_server.controllers.offers_controller
    post:
      tags:
      - Offers
      description: Add offers
      operationId: add_offer
      requestBody:
        description: Adding offers
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/offer_info'
      responses:
        "201":
          description: New offer added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/201OfferCreatedResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/409ConflictResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
      x-openapi-router-controller: swagger_server.controllers.offers_controller
  /offers/{offer_id}:
    delete:
      tags:
      - Offers
      description: delete offers
      operationId: delete_offer
      parameters:
      - name: offer_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: offer deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200OfferDeletedResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/409ConflictResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
      x-openapi-router-controller: swagger_server.controllers.offers_controller
    patch:
      tags:
      - Offers
      description: update offers
      operationId: update_offers
      parameters:
      - name: offer_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Adds price and offers
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/update_offer_info'
      responses:
        "200":
          description: new fields added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200UpdateOfferResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/409ConflictResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
      x-openapi-router-controller: swagger_server.controllers.offers_controller
  /prices:
    get:
      tags:
      - Prices
      description: Get prices
      operationId: get_prices
      responses:
        "200":
          description: Get all price
          content:
            applications/json:
              schema:
                $ref: '#/components/schemas/price_details'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404NotFoundResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
        default:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/400BadRequestResponse'
      x-openapi-router-controller: swagger_server.controllers.prices_controller
    post:
      tags:
      - Prices
      description: Add price to cars
      operationId: add_price
      requestBody:
        description: Adding price
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/price_info'
      responses:
        "200":
          description: 200 responses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/201PriceCreatedResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404NotFoundResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
        default:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/400BadRequestResponse'
      x-openapi-router-controller: swagger_server.controllers.prices_controller
  /price/{price_id}:
    get:
      tags:
      - Prices
      description: get unique id
      operationId: get_price
      parameters:
      - name: price_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Fetch prices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/price_details'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404NotFoundResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
        default:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/400BadRequestResponse'
      x-openapi-router-controller: swagger_server.controllers.prices_controller
    delete:
      tags:
      - Prices
      description: Delete price
      operationId: delete_price
      parameters:
      - name: price_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: user deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200PriceDeletedResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/401UnauthorizedResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/409ConflictResponse'
        "503":
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/503ServerUnavailableResponse'
      x-openapi-router-controller: swagger_server.controllers.prices_controller
components:
  schemas:
    carmodeldetails:
      allOf:
      - $ref: '#/components/schemas/carmodelinfo'
      - type: object
        properties:
          model_id:
            type: string
    carmodelinfo:
      required:
      - fuel_type
      - model_name
      - model_type
      - transmission
      - trim
      type: object
      properties:
        model_type:
          type: string
          enum:
          - SUVs
          - Hatchback
          - Sedan
          - pickuptruck
          - MUV
          - off road
        model_name:
          type: string
        transmission:
          type: string
          enum:
          - manual
          - automatic
        fuel_type:
          type: string
          enum:
          - diesel
          - petrol
          - ev
          - cng
        trim:
          type: string
          enum:
          - SE
          - DX
          - GT
    "201CarModelCreatedResponse":
      type: object
      properties:
        code:
          type: integer
          default: 201
        message:
          type: string
          default: Car Model created
        data:
          $ref: '#/components/schemas/carmodeldetails'
      example:
        code: 0
        data: ""
        message: Car Model created
    offer_info:
      required:
      - applicable_models
      - discount_pct
      - end_date
      - start_date
      type: object
      properties:
        discount_pct:
          type: number
        start_date:
          type: string
        end_date:
          type: string
        applicable_models:
          type: array
          items:
            type: string
      example:
        end_date: 
        discount_pct: 6.02745618307040320615897144307382404804229736328125
        applicable_models:
        - applicable_models
        - applicable_models
        start_date: 
    offer_details:
      allOf:
      - $ref: '#/components/schemas/offer_info'
      - type: object
        properties:
          offer_id:
            type: string
    update_offer_info:
      type: object
      properties:
        key:
          type: string
          enum:
          - discount_pct
          - start_date
          - end_date
          - applicable_models
        value:
          type: string
    "201OfferCreatedResponse":
      type: object
      properties:
        code:
          type: integer
          default: 201
        message:
          type: string
          default: Offer created
        data:
          $ref: '#/components/schemas/offer_details'
      example:
        code: 0
        data: ""
        message: Offer created
    "200UpdateOfferResponse":
      type: object
      properties:
        code:
          type: integer
          default: 200
        message:
          type: string
          default: ok
        data:
          type: array
          items:
            $ref: '#/components/schemas/offer_info'
      example:
        code: 0
        data:
        - end_date: 5.962133916683182377482808078639209270477294921875
          discount_pct: 6.02745618307040320615897144307382404804229736328125
          applicable_models:
          - applicable_models
          - applicable_models
          start_date: 1.46581298050294517310021547018550336360931396484375
        - end_date: 5.962133916683182377482808078639209270477294921875
          discount_pct: 6.02745618307040320615897144307382404804229736328125
          applicable_models:
          - applicable_models
          - applicable_models
          start_date: 1.46581298050294517310021547018550336360931396484375
        message: ok
    "200OfferDeletedResponse":
      type: object
      properties:
        code:
          type: integer
          default: 200
        message:
          type: string
          default: ok
        data:
          type: array
          items:
            $ref: '#/components/schemas/offer_details'
      example:
        code: 0
        data:
        - ""
        - ""
        message: ok
    price_info:
      required:
      - end_date
      - model_id
      - price
      - start_date
      type: object
      properties:
        model_id:
          type: string
        price:
          type: number
        start_date:
          type: string
        end_date:
          type: string
    price_details:
      allOf:
      - $ref: '#/components/schemas/price_info'
      - type: object
        properties:
          price_id:
            type: string
    "201PriceCreatedResponse":
      type: object
      properties:
        code:
          type: integer
          default: 201
        message:
          type: string
          default: ok
        data:
          type: array
          items:
            $ref: '#/components/schemas/price_details'
      example:
        code: 0
        data:
        - ""
        - ""
        message: ok
    "200PriceDeletedResponse":
      type: object
      properties:
        code:
          type: integer
          default: 200
        message:
          type: string
          default: ok
        data:
          type: array
          items:
            $ref: '#/components/schemas/price_details'
      example:
        code: 0
        data:
        - ""
        - ""
        message: ok
    "400BadRequestResponse":
      type: object
      properties:
        code:
          type: integer
          default: 400
        message:
          type: string
          default: BAD_REQUEST
        data:
          type: array
          items:
            type: object
    "401UnauthorizedResponse":
      type: object
      properties:
        code:
          type: integer
          default: 401
        message:
          type: string
          default: UNAUTHORIZED
    "403ForbiddenResponse":
      type: object
      properties:
        code:
          type: integer
          default: 403
        message:
          type: string
          default: FORBIDDEN
    "404NotFoundResponse":
      type: object
      properties:
        code:
          type: integer
          default: 404
        message:
          type: string
          default: NOT_FOUND
    "503ServerUnavailableResponse":
      type: object
      properties:
        code:
          type: integer
          default: 503
        message:
          type: string
          default: SERVER_UNAVAILABLE
    "409ConflictResponse":
      type: object
      properties:
        code:
          type: integer
          default: 409
        message:
          type: string
          default: CONFLICT